+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.0.0] - 2026-08-31
+
+_Stable release based on [6.0.0-rc.1]._
+
 ## [6.0.0-rc.1] - 2026-08-31
 
 ### Added
@@ -1874,6 +1878,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[6.0.0]: https://https://github.com/internetguru/laravel-common/compare/v5.27.1...v6.0.0
 [6.0.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.27.1
 [5.27.1]: https://https://github.com/internetguru/laravel-common/compare/v5.27.0...v5.27.1
 [5.27.0]: https://https://github.com/internetguru/laravel-common/compare/v5.26.0...v5.27.0